% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_diff.R
\name{cal_diff}
\alias{cal_diff}
\title{DiffDynpkm}
\usage{
cal_diff(treat, control, method = c("glm", "wilcox")[1], shrink = FALSE)
}
\arguments{
\item{treat}{need a dataframe,treatment genes expression}

\item{control}{need a dataframe,control genes expression}

\item{method}{'glm' or 'wilcox' calculate p value}

\item{shrink}{NB model shrink}
}
\value{
dataframe
}
\description{
Calculate the differential expression of genes using fpkm/rpkm/tpm/cpm
}
\examples{
data('gene_fpkm')
treat = gene_fpkm[,1:3]
control = gene_fpkm[,4:6]
cal_diff(treat,control)
}
