% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_dyn.R
\name{cal_dyn}
\alias{cal_dyn}
\title{DiffDynpkm}
\usage{
cal_dyn(design, exp, gene = "gene_1", k = c(3, 3), bs = "cr")
}
\arguments{
\item{design}{need a dataframe,expriment matrix}

\item{exp}{need a dataframe,genes expression matrix}

\item{gene}{target gene}

\item{k}{the dimension of the basis used to represent the smooth term}

\item{bs}{a two letter character string indicating the (penalized) smoothing basis to use. (eg "tp" for thin plate regression spline, "cr" for cubic regression spline)}
}
\value{
list
}
\description{
Predic genes dynamic expression in different conditions
}
